Fixing nodes.pp with hiera('variable') calls
So... either this is a new bug, or I failed at noticing it. Everything
seemed to be working properly except package installs, but I think
that's because ssh keys had been installed previously through the old
person module. Apparently, in nodes.pp (since it's not a class), I need
to explicitly assign hiera set variables with the hiera function call,
and not to a default empty variable ([] or {} as the case may be),
which *does* work in classes.